+++ AI/AI_test/test.cpp
@@ -1,16 +1,14 @@
 #include "pch.h"
 #include <GridFunctions.h>
 
-
-
-
-
-TEST(Grid, ComponentCreation) {
+TEST(GridFunctions, InitGrid) {
+  
   ecs::EntityComponentSystem mEcs;
-  GridFunctions::InitGrid(mEcs, 9, 9, 4);  
-  //  GridFunctions::CreateDebugSystems(mEcs);
+  GridFunctions::InitGrid(mEcs, 50, 50, 4);  
+  GridFunctions::CreateDebugSystems(mEcs);
 
-  //  mEcs.update(0);
+  mEcs.update(0);
+
   EXPECT_EQ(1, 1);
   EXPECT_TRUE(true);
 }
